archive/requirements.txt
# LangGraph and LangChain
langgraph==0.2.34
langchain==0.3.6
langchain-community==0.3.4
langchain-core==0.3.15

# LLM Integration
ollama==0.3.3

# Notion API
notion-client==2.2.1

# MCP Protocol
mcp==1.21.0

# Streamlit UI
streamlit==1.39.0
streamlit-chat==0.1.1

# Utilities
python-dotenv==1.0.0
pydantic==2.11.0
requests==2.32.3
aiohttp==3.10.5

# Data handling
pandas==2.2.3
